Crate re_renderer

Rerun’s renderer.

A wgpu based renderer wgpu for all your visualization needs. Used in re_runner to display the contents of any view contents other than pure UI.

Feature flags

arrow (enabled by default) — Support for Arrow datatypes for end-to-end zero-copy.

import-obj (enabled by default) — Support importing .obj files

import-gltf (enabled by default) — Support importing .gltf and .glb files

serde — Enable (de)serialization using serde.

Re-exports

pub use view_builder::AutoSizeConfig;

Modules

allocator 🔒

High level GPU memory allocators.

color 🔒

colormap 🔒

config

context 🔒

debug_label 🔒

depth_offset 🔒

draw_phases 🔒

error_tracker 🔒

Special error handling datastructures for debug builds (never crash!).

file_resolver 🔒

This module implements one half of our cross-platform #import system.

file_server 🔒

file_system 🔒

global_bindings 🔒

importer

line_strip_builder 🔒

mesh

point_cloud_builder 🔒

rect 🔒

renderer

resource_managers

Resource managers are concerned with mapping (typically) higher level user data to their Gpu representation.

size 🔒

view_builder

wgpu_buffer_types 🔒

Explicitly padded and/or aligned types following wgsl rules. See wgsl spec on alignment and size

wgpu_resources 🔒

Wgpu resource pools are concerned with handling low level gpu resources efficiently.

Macros

include_file

A macro to read the contents of a file on disk, and resolve #import clauses as required.

include_shader_module

Create a shader module using the include_file! macro and set the path name as debug string.

Structs

Color32

This format is used for space-efficient color representation (32 bits).

DebugLabel

Label for resources. Optimized out in release builds.

FileResolver

The FileResolver handles both resolving import clauses and doing the actual string interpolation.

FileServer

A file server capable of watching filesystem events in the background and resolve #import clauses in files.

ImportClause

A pre-parsed import clause, as in #import <something>.

IntRect

A 2D rectangle with integer coordinates.

LineStripBuilder

LineStripSeriesBuilder

Builder for a vector of line strips, making it easy to create crate::renderer::LineDrawData.

OutlineConfig

OutlineMaskPreference

What outline (if any) should be drawn.

PickingLayerId

Combination of PickingLayerObjectId and PickingLayerInstanceId.

PickingLayerInstanceId

The second 64bit of the picking layer.

PickingLayerObjectId

The first 64bit of the picking layer.

PickingLayerProcessor

Manages the rendering of the picking layer pass, its render targets & readback buffer.

PointCloudBatchBuilder

PointCloudBuilder

Builder for point clouds, making it easy to create crate::renderer::PointCloudDrawData.

RenderContext

Any resource involving wgpu rendering which can be re-used across different scenes. I.e. render pipelines, resource pools, etc.

Rgba

0-1 linear space RGBA color with premultiplied alpha.

Rgba32Unmul

RGBA color in sRGB gamma space, with separate/unmultiplied linear alpha.

ScreenshotProcessor

SearchPath

Specifies where to look for imports when both absolute and relative resolution fail.

Size

A size of something in either world-units, screen-units, or unsized.

WgpuResourcePoolStatistics

Enums

ColorMap

Traits

FileSystem

A very limited filesystem, just enough for our internal needs.

Functions

colormap_inferno_srgb

Returns sRGB polynomial approximation from Inferno color map, assuming t is normalized.

colormap_magma_srgb

Returns sRGB polynomial approximation from Magma color map, assuming t is normalized.

colormap_plasma_srgb

Returns sRGB polynomial approximation from Plasma color map, assuming t is normalized.

colormap_srgb

colormap_turbo_srgb

Returns sRGB polynomial approximation from Turbo color map, assuming t is normalized.

colormap_viridis_srgb

Returns sRGB polynomial approximation from Viridis color map, assuming t is normalized.

get_filesystem

Returns the recommended filesystem handle for the current platform.

grayscale_srgb

Returns an sRGB gray value, assuming t is normalized.

new_recommended_file_resolver

Returns the recommended FileResolver for the current platform/target.

Type Definitions

Buffer

DepthOffset

Depth offset used to resolve z-fighting between 2d primitives.

GpuReadbackIdentifier

Identifier used to identify a buffer upon retrieval of the data.

RecommendedFileResolver

The recommended FileResolver type for the current platform/target.