Module re_renderer::colormap

Enums

Colormap

Constants

B2 ğŸ”’

B4 ğŸ”’

C0 ğŸ”’

C0 ğŸ”’

C0 ğŸ”’

C0 ğŸ”’

C1 ğŸ”’

C1 ğŸ”’

C1 ğŸ”’

C1 ğŸ”’

C2 ğŸ”’

C2 ğŸ”’

C2 ğŸ”’

C2 ğŸ”’

C3 ğŸ”’

C3 ğŸ”’

C3 ğŸ”’

C3 ğŸ”’

C4 ğŸ”’

C4 ğŸ”’

C4 ğŸ”’

C4 ğŸ”’

C5 ğŸ”’

C5 ğŸ”’

C5 ğŸ”’

C5 ğŸ”’

C6 ğŸ”’

C6 ğŸ”’

C6 ğŸ”’

C6 ğŸ”’

G2 ğŸ”’

G4 ğŸ”’

R2 ğŸ”’

R4 ğŸ”’

Functions

colormap_inferno_srgb

Returns sRGB polynomial approximation from Inferno color map, assuming t is normalized.

colormap_magma_srgb

Returns sRGB polynomial approximation from Magma color map, assuming t is normalized.

colormap_plasma_srgb

Returns sRGB polynomial approximation from Plasma color map, assuming t is normalized.

colormap_srgb

colormap_turbo_srgb

Returns sRGB polynomial approximation from Turbo color map, assuming t is normalized.

colormap_viridis_srgb

Returns sRGB polynomial approximation from Viridis color map, assuming t is normalized.

grayscale_srgb

Returns an sRGB gray value, assuming t is normalized.