Struct re_renderer::draw_phases::picking_layer::DepthReadbackWorkaround

struct DepthReadbackWorkaround {
    render_pipeline: GpuRenderPipelineHandle,
    bind_group: GpuBindGroup,
    readable_texture: Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>,
}

Utility for copying a depth texture when it can’t be read-back directly to a [wgpu::TextureFormat::R32Float] which is readable texture.

Implementation note: This is a plain & simple “sample in shader and write to texture” utility. It might be worth abstracting this further into a general purpose operator. There is not much in here that is specific to the depth usecase!

Fields
render_pipeline: GpuRenderPipelineHandlebind_group: GpuBindGroupreadable_texture: Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>
Implementations

impl DepthReadbackWorkaround

const READBACK_FORMAT: TextureFormat = wgpu::TextureFormat::Rgba32Float

There’s two layers of workarounds here:

WebGL (via spec) not being able to read back depth textures

unclear behavior for any readback that isn’t RGBA Furthermore, integer textures also seemed to be problematic, but it seems to work fine for [wgpu::TextureFormat::Rgba32Uint] which we use for our picking ID Details see wgpu#3644

fn new( ctx: &mut RenderContext, extent: UVec2, depth_target_handle: GpuTextureHandle ) -> DepthReadbackWorkaround

fn copy_to_readable_texture( &self, encoder: &mut CommandEncoder, pools: &WgpuResourcePools, global_binding_bind_group: &GpuBindGroup ) -> Result<&Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>, PoolError>