re_renderer::draw_phases
Module screenshot

Easy screenshot taking.

Executes compositing into a special target if we want to take a screenshot. This is necessary because composite is expected to write directly into the final output from which we can’t read back and of which we don’t control the format.

This comes with the perk that we can do extra things here if we want!

TODO(andreas): One thing to add would be more anti-aliasing! We could render the same image with subpixel moved camera in order to get super-sampling without hitting texture size limitations. Or alternatively try to render the images in several tiles 🤔. In any case this would greatly improve quality!

Structs

ReadbackBeltMetadata 🔒

Type used as user data on the gpu readback belt.

ScreenshotProcessor