Module re_renderer::error_handling::now_or_never

Future utility copied from Bevy https://github.com/bevyengine/bevy/blob/4b1865f8bda3c3d6b3f399fb6727635b3ffcbb41/crates/bevy_utils/src/futures.rs

It is used there for a very similar purpose: catching errors on native wgpu which are known to be non-asynchronous.

Constants

NOOP_WAKER_VTABLE 🔒

Functions

noop 🔒 ⚠

noop_clone 🔒 ⚠

noop_raw_waker 🔒

noop_waker 🔒

now_or_never

Consumes the future, polls it once, and immediately returns the output or returns None if it wasn’t ready yet.