Module re_renderer::error_handling::wgpu_core_error

Macros

impl_trait 🔒

E.g. to implement DedupableError for u32 + u64:

try_downcast 🔒

Tries downcasting a given value into the specified possibilities (or all of them if none are specified), then run the given expression on the downcasted value.

Structs

WrappedContextError

A wgpu_core::ContextError with hashing and equality capabilities.

Traits

DedupableError 🔒

An error with some extra deduplication logic baked in.

Functions

type_of_var 🔒