Struct re_renderer::file_server::file_server_impl::FileServer

pub struct FileServer {
    watcher: RecommendedWatcher,
    events_rx: Receiver<Event>,
    file_watch_count: HashMap<PathBuf, usize>,
}

A file server capable of watching filesystem events in the background and resolve #import clauses in files.

Fields
watcher: RecommendedWatcherevents_rx: Receiver<Event>file_watch_count: HashMap<PathBuf, usize>
Implementations

impl FileServer

fn new() -> Result<Self>

impl FileServer

pub fn get<T>(f: impl FnMut(&FileServer) -> T) -> T

Returns a reference to the global FileServer.

pub fn get_mut<T>(f: impl FnMut(&mut FileServer) -> T) -> T

Returns a mutable reference to the global FileServer.

impl FileServer

pub fn watch<Fs: FileSystem>( &mut self, resolver: &mut FileResolver<Fs>, path: impl AsRef<Path>, recursive: bool ) -> Result<PathBuf>

Starts watching for file events at the given path.

The given path is canonicalized.

pub fn unwatch(&mut self, path: impl AsRef<Path>) -> Result<()>

Stops watching for file events at the given path.

pub fn collect<Fs: FileSystem>( &mut self, resolver: &mut FileResolver<Fs> ) -> HashSet<PathBuf>

Coalesces all filesystem events since the last call to collect, and returns a set of all modified paths.