Module re_renderer::file_system

Structs

MemFileSystem

A FileSystem implementation backed by an hash map.

OsFileSystem

A FileSystem implementation that simply delegates to std::fs.

Statics

MEM_FILE_SYSTEM 🔒

The global MemFileSystem.

Traits

FileSystem

A very limited filesystem, just enough for our internal needs.

Functions

get_filesystem

Returns the recommended filesystem handle for the current platform.