re_renderer::importer
Struct CpuModel

pub struct CpuModel {
    pub meshes: SlotMap<CpuModelMeshKey, CpuMesh>,
    pub instances: Vec<CpuMeshInstance>,
}

A collection of meshes & mesh instances on the CPU.

Note that there is currently no GpuModel equivalent, since GpuMeshInstancees use shared ownership of GpuMeshes.

This is the output of a model loader and is ready to be converted into a series of GpuMeshInstances that can be rendered.

This is meant as a useful intermediate structure for doing post-processing steps on the model prior to gpu upload.

Fields
meshes: SlotMap<CpuModelMeshKey, CpuMesh>instances: Vec<CpuMeshInstance>
Implementations

impl CpuModel

pub fn from_single_mesh(mesh: CpuMesh) -> Self

Creates a new CpuModel from a single CpuMesh, creating a single instance with identity transform.

pub fn add_single_instance_mesh(&mut self, mesh: CpuMesh)

Adds a new CpuMesh to the model, creating a single instance with identity transform.

pub fn calculate_bounding_box(&self) -> BoundingBox

pub fn into_gpu_meshes( self, ctx: &RenderContext, ) -> Result<Vec<GpuMeshInstance>, MeshError>

Converts the entire model into a serious of mesh instances that can be rendered.

Silently ignores:

instances with invalid mesh keys

unreferenced meshes