Struct re_renderer::line_drawable_builder::LineDrawableBuilder

pub struct LineDrawableBuilder<'ctx> {
    pub ctx: &'ctx RenderContext,
    pub(crate) vertices_buffer: DataTextureSource<'ctx, LineVertex>,
    pub(crate) batches: Vec<LineBatchInfo>,
    pub(crate) strips_buffer: DataTextureSource<'ctx, LineStripInfo>,
    pub(crate) picking_instance_ids_buffer: DataTextureSource<'ctx, PickingLayerInstanceId>,
    pub(crate) radius_boost_in_ui_points_for_outlines: f32,
}

Builder for a vector of line strips, making it easy to create crate::renderer::LineDrawData.

TODO(andreas): We could make significant optimizations here by making this builder capable of writing to a GPU readable memory location for all its data.

Fields
ctx: &'ctx RenderContextvertices_buffer: DataTextureSource<'ctx, LineVertex>batches: Vec<LineBatchInfo>strips_buffer: DataTextureSource<'ctx, LineStripInfo>picking_instance_ids_buffer: DataTextureSource<'ctx, PickingLayerInstanceId>

Buffer for picking instance id - every strip gets its own instance id. Therefore, there need to be always as many picking instance ids as there are strips.

radius_boost_in_ui_points_for_outlines: f32
Implementations

impl<'ctx> LineDrawableBuilder<'ctx>

pub fn new(ctx: &'ctx RenderContext) -> Self

pub fn reserve_strips( &mut self, num_strips: usize ) -> Result<usize, CpuWriteGpuReadError>

Returns number of strips that can be added without reallocation. This may be smaller than the requested number if the maximum number of strips is reached.

pub fn reserve_vertices( &mut self, num_vertices: usize ) -> Result<usize, CpuWriteGpuReadError>

Returns number of vertices that can be added without reallocation. This may be smaller than the requested number if the maximum number of vertices is reached.

pub fn radius_boost_in_ui_points_for_outlines( &mut self, radius_boost_in_ui_points_for_outlines: f32 )

Boosts the size of the points by the given amount of ui-points for the purpose of drawing outlines.

pub fn batch( &mut self, label: impl Into<DebugLabel> ) -> LineBatchBuilder<'_, 'ctx>

Start of a new batch.

pub fn into_draw_data(self) -> Result<LineDrawData, LineDrawDataError>

Finalizes the builder and returns a line draw data with all the lines added so far.

pub fn is_empty(&self) -> bool

pub fn default_box_flags() -> LineStripFlags