Module re_renderer::renderer

Modules

compositor 🔒

debug_overlay 🔒

depth_cloud 🔒

Renderer that makes it easy to draw point clouds straight out of depth textures.

generic_skybox 🔒

lines 🔒

Line renderer for efficient rendering of many line(strips)

mesh_renderer 🔒

Mesh renderer.

point_cloud 🔒

Point renderer for efficient rendering of point clouds.

rectangles 🔒

Renderer that makes it easy to draw textured 2d rectangles with transparency

test_triangle 🔒

Structs

DebugOverlayDrawData

Debug overlay for quick & dirty display of texture contents.

DebugOverlayRenderer

DepthCloud

DepthCloudDrawData

DepthCloudRenderer

DepthClouds

GenericSkyboxDrawData

LineBatchInfo

Data that is valid for a batch of line strips.

LineDrawData

A line drawing operation. Encompasses several lines, each consisting of a list of positions. Expected to be recrated every frame.

LineStripFlags

Property flags for a line strip

LineStripInfo

Style information for a line strip.

LineVertex

MeshDrawData

MeshInstance

PointCloudBatchFlags

Property flags for a point batch

PointCloudBatchInfo

Data that is valid for a batch of point cloud points.

PointCloudDrawData

A point cloud drawing operation. Expected to be recrated every frame.

PointCloudVertex

Description of a point cloud.

RectangleDrawData

TestTriangleDrawData

TexturedRect

Enums

DepthCloudDepthData

The raw data from a depth texture.

PointCloudDrawDataError

TextureFilterMag

Texture filter setting for magnification (a texel covers several pixels).

TextureFilterMin

Texture filter setting for minification (several texels fall to one pixel).

Traits

DrawData

GPU sided data used by a Renderer to draw things to the screen.

Renderer

A Renderer encapsulate the knowledge of how to render a certain kind of primitives.

Functions

screen_triangle_vertex_shader

Gets or creates a vertex shader module for drawing a screen filling triangle.