re_renderer::renderer::rectangles
Struct ColormappedTexture

pub struct ColormappedTexture {
    pub texture: GpuTexture2D,
    pub range: [f32; 2],
    pub decode_srgb: bool,
    pub multiply_rgb_with_alpha: bool,
    pub gamma: f32,
    pub color_mapper: ColorMapper,
    pub shader_decoding: Option<ShaderDecoding>,
}

Describes a texture and how to map it to a color.

Fields
texture: GpuTexture2Drange: [f32; 2]

Min/max range of the values in the texture.

Used to normalize the input values (squash them to the 0-1 range). The normalization is applied before sRGB gamma decoding and alpha pre-multiplication (this transformation is also applied to alpha!).

decode_srgb: bool

Decode 0-1 sRGB gamma values to linear space before filtering?

Only applies to [wgpu::TextureFormat::Rgba8Unorm] and float textures.

multiply_rgb_with_alpha: bool

Multiply color channels with the alpha channel before filtering?

Set this to false for textures that donโt have an alpha channel or are already pre-multiplied. Applied after range normalization and srgb decoding, before filtering.

gamma: f32

Raise the normalized values to this power (before any color mapping). Acts like an inverse brightness.

Default: 1.0

color_mapper: ColorMapper

For any one-component texture, you need to supply a color mapper, which maps the normalized .r component to a color.

Setting a color mapper for a four-component texture is an error. Failure to set a color mapper for a one-component texture is an error.

shader_decoding: Option<ShaderDecoding>

For textures that need decoding in the shader, for example NV12 encoded images.

Implementations

impl ColormappedTexture

pub fn from_unorm_rgba(texture: GpuTexture2D) -> Self

Assumes a separate/unmultiplied alpha.

pub fn width_height(&self) -> [u32; 2]