Struct re_renderer::renderer::DepthCloud

pub struct DepthCloud {
    pub world_from_rdf: Affine3A,
    pub depth_camera_intrinsics: Mat3,
    pub world_depth_from_texture_depth: f32,
    pub point_radius_from_world_depth: f32,
    pub max_depth_in_world: f32,
    pub depth_dimensions: UVec2,
    pub depth_texture: GpuTexture2D,
    pub colormap: Colormap,
    pub outline_mask_id: OutlineMaskPreference,
    pub picking_object_id: PickingLayerObjectId,
}

Fields
world_from_rdf: Affine3A

The extrinsics of the camera used for the projection, with a RDF coordinate system on the right-hand side.

depth_camera_intrinsics: Mat3

The intrinsics of the camera used for the projection.

Only supports pinhole cameras at the moment.

world_depth_from_texture_depth: f32

Multiplier to get world-space depth from whatever is in Self::depth_texture.

point_radius_from_world_depth: f32

Point radius is calculated as world-space depth times this value.

max_depth_in_world: f32

The maximum depth value in world-space, for use with the colormap.

depth_dimensions: UVec2

The dimensions of the depth texture in pixels.

depth_texture: GpuTexture2D

The actual data from the depth texture.

Only textures with sample type Float are supported.

colormap: Colormap

Configures color mapping mode.

outline_mask_id: OutlineMaskPreference

Option outline mask id preference.

picking_object_id: PickingLayerObjectId

Picking object id that applies for the entire depth cloud.

Implementations

impl DepthCloud

pub fn bbox(&self) -> BoundingBox

World-space bounding-box.