re_renderer::renderer
Trait DrawData

pub trait DrawData {
    type Renderer: Renderer<RendererDrawData = Self> + Send + Sync;
}

GPU sided data used by a Renderer to draw things to the screen.

Valid only for the frame in which it was created (typically uses temp allocations!).

Required Associated Types

type Renderer: Renderer<RendererDrawData = Self> + Send + Sync