Module re_renderer::resource_managers

Resource managers are concerned with mapping (typically) higher level user data to their Gpu representation.

They facilitate fast & easy gpu upload and resource usage.

This is in contrast to the pools in crate::wgpu_resources which are exclusively concerned with low level gpu resources and their efficient allocation.

Modules

image_data_to_texture 🔒

For an overview of image data interpretation check re_video’s decoder docs!

texture_manager 🔒

yuv_converter 🔒

Structs

GpuTexture2D

Handle to a 2D resource.

ImageDataDesc

Describes image data for the purpose of creating a 2D texture.

TextureManager2D

Texture manager for 2D textures.

Enums

ImageDataToTextureError

Error that can occur when converting image data to a texture.

SourceImageDataFormat

Image data format that can be converted to a wgpu texture.

TextureManager2DError

YuvMatrixCoefficients

Yuv matrix coefficients that determine how a YUV image is meant to be converted to RGB.

YuvPixelLayout

Supported chroma subsampling input formats.

YuvRange

Expected range of YUV values.

Functions

transfer_image_data_to_texture

Takes raw image data and transfers & converts it to a GPU texture.