re_renderer::resource_managers
Module image_data_to_texture

For an overview of image data interpretation check re_video’s decoder docs!

Structs

ImageDataDesc

Describes image data for the purpose of creating a 2D texture.

Enums

ImageDataToTextureError

Error that can occur when converting image data to a texture.

SourceImageDataFormat

Image data format that can be converted to a wgpu texture.

Functions

copy_data_to_texture 🔒

transfer_image_data_to_texture

Takes raw image data and transfers & converts it to a GPU texture.