re_renderer::resource_managers::image_data_to_texture
Enum SourceImageDataFormat

pub enum SourceImageDataFormat {
    WgpuCompatible(TextureFormat),
    Yuv {
        layout: YuvPixelLayout,
        coefficients: YuvMatrixCoefficients,
        range: YuvRange,
    },
}

Image data format that can be converted to a wgpu texture.

Variants

WgpuCompatible(TextureFormat)

The source format is already in a wgpu compatible format.

โ๏ธ Only because a format is listed in wgpu::TextureFormat doesnโt mean we can use it on the currently active backend. TODO(andreas): This is a temporary measure until we cover what rerun covers. Weโd really like incoming data to not reason with [wgpu::TextureFormat] since itโs so hard to know whatโs appropriate & available for a given device.

Yuv

YUV (== YCbCr) formats, typically using chroma downsampling.

Does not handle chroma sample locations.

Fields

layout: YuvPixelLayout

coefficients: YuvMatrixCoefficients

range: YuvRange