Struct re_renderer::resource_managers::TextureManager2D

pub struct TextureManager2D {
    white_texture_unorm: GpuTexture2DHandle,
    zeroed_texture_uint: GpuTexture2DHandle,
    device: Arc<Device>,
    queue: Arc<Queue>,
    texture_cache: HashMap<u64, GpuTexture2DHandle>,
    accessed_textures: HashSet<u64>,
}

Texture manager for 2D textures.

The scope is intentionally limited to particular kinds of textures that currently require this kind of handle abstraction/management. More complex textures types are typically handled within renderer which utilize the texture pool directly. This manager in contrast, deals with user provided texture data! We might revisit this later and make this texture manager more general purpose.

Fields
white_texture_unorm: GpuTexture2DHandlezeroed_texture_uint: GpuTexture2DHandledevice: Arc<Device>queue: Arc<Queue>texture_cache: HashMap<u64, GpuTexture2DHandle>accessed_textures: HashSet<u64>
Implementations

impl TextureManager2D

pub(crate) fn new(
device: Arc<Device>,
queue: Arc<Queue>,
texture_pool: &mut GpuTexturePool
) -> Self

pub fn create(
&mut self,
texture_pool: &mut GpuTexturePool,
creation_desc: &Texture2DCreationDesc<'_>
) -> GpuTexture2DHandle

Creates a new 2D texture resource and schedules data upload to the GPU. TODO(jleibs): All usages of this should be be replaced with get_or_create, which is strictly preferable

pub fn get_or_create(
&mut self,
key: u64,
texture_pool: &mut GpuTexturePool,
creation_desc: &Texture2DCreationDesc<'_>
) -> GpuTexture2DHandle

Creates a new 2D texture resource and schedules data upload to the GPU if a texture wasn’t already created using the same key.

pub fn white_texture_unorm_handle(&self) -> &GpuTexture2DHandle

Returns a single pixel white pixel with an rgba8unorm format.

pub fn white_texture_unorm(
&self
) -> &Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>

Returns a single pixel white pixel with an rgba8unorm format.

pub fn zeroed_texture_uint(
&self
) -> &Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>

Returns a single pixel white pixel with an rgba8unorm format.

fn create_and_upload_texture(
device: &Device,
queue: &Queue,
texture_pool: &mut GpuTexturePool,
creation_desc: &Texture2DCreationDesc<'_>
) -> GpuTexture2DHandle

pub(crate) fn get(
&self,
handle: &GpuTexture2DHandle
) -> Result<Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>, ResourceManagerError>

Retrieves gpu handle.

pub(crate) fn begin_frame(&mut self, _frame_index: u64)