Type Alias re_renderer::resource_managers::GpuMeshHandle

pub type GpuMeshHandle = ResourceHandle<MeshHandleInner>;

Aliased Type

enum GpuMeshHandle {
    LongLived(Arc<MeshHandleInner, Global>),
    Frame {
        key: MeshHandleInner,
        valid_frame_index: u64,
    },
    Invalid,
}

Variants

LongLived(Arc<MeshHandleInner, Global>)

Resource handle that keeps the resource alive as long as there are handles.

Once the user drops the last handle, the resource will be discarded on next frame maintenance. (actual reclaiming of GPU resources may take longer)

Frame

Fields

key: MeshHandleInner

valid_frame_index: u64

This handle is only valid for this frame. Querying it during any other frame will fail.

Handle that is valid for a single frame

Invalid

No handle, causes error on resolve.