Struct re_renderer::LineStripSeriesBuilder

pub struct LineStripSeriesBuilder {
    pub vertices: Vec<LineVertex>,
    pub batches: Vec<LineBatchInfo>,
    pub strips: Vec<LineStripInfo>,
    pub(crate) picking_instance_ids_buffer: CpuWriteGpuReadBuffer<PickingLayerInstanceId>,
    pub(crate) radius_boost_in_ui_points_for_outlines: f32,
}

Builder for a vector of line strips, making it easy to create crate::renderer::LineDrawData.

TODO(andreas): We could make significant optimizations here by making this builder capable of writing to a GPU readable memory location. This will require some ahead of time size limit, but should be feasible. But before that we first need to sort out cpu->gpu transfers better by providing staging buffers.

Fields
vertices: Vec<LineVertex>batches: Vec<LineBatchInfo>strips: Vec<LineStripInfo>picking_instance_ids_buffer: CpuWriteGpuReadBuffer<PickingLayerInstanceId>

Buffer for picking instance id - every strip gets its own instance id. Therefore, there need to be always as many picking instance ids as there are strips.

radius_boost_in_ui_points_for_outlines: f32
Implementations

impl LineStripSeriesBuilder

pub fn new(ctx: &RenderContext) -> Self

pub fn radius_boost_in_ui_points_for_outlines( self, radius_boost_in_ui_points_for_outlines: f32 ) -> Self

Boosts the size of the points by the given amount of ui-points for the purpose of drawing outlines.

pub fn batch(&mut self, label: impl Into<DebugLabel>) -> LineBatchBuilder<'_>

Start of a new batch.

pub fn iter_vertices_by_batch( &self ) -> impl Iterator<Item = (&LineBatchInfo, impl Iterator<Item = &LineVertex>)>

pub fn into_draw_data( self, ctx: &mut RenderContext ) -> Result<LineDrawData, LineDrawDataError>

Finalizes the builder and returns a line draw data with all the lines added so far.

pub fn is_empty(&self) -> bool

pub fn default_box_flags() -> LineStripFlags