Struct re_renderer::PointCloudBatchBuilder

pub struct PointCloudBatchBuilder<'a>(&'a mut PointCloudBuilder);

Tuple Fields
0: &'a mut PointCloudBuilder
Implementations

impl<'a> PointCloudBatchBuilder<'a>

fn batch_mut(&mut self) -> &mut PointCloudBatchInfo

pub fn world_from_obj(self, world_from_obj: Affine3A) -> Self

Sets the world_from_obj matrix for the entire batch.

pub fn outline_mask_ids(self, outline_mask_ids: OutlineMaskPreference) -> Self

Sets an outline mask for every element in the batch.

pub fn depth_offset(self, depth_offset: DepthOffset) -> Self

Sets the depth offset for the entire batch.

pub fn add_points( self, positions: &[Vec3], radii: &[Size], colors: &[Color32], picking_ids: &[PickingLayerInstanceId] ) -> Self

Add several 3D points

Returns a PointBuilder which can be used to set the colors, radii, and user-data for the points.

Will add all positions. Missing radii will default to Size::AUTO. Missing colors will default to white.

pub fn add_points_2d( self, positions: &[Vec3], radii: &[Size], colors: &[Color32], picking_ids: &[PickingLayerInstanceId] ) -> Self

Adds several 2D points (assumes Z=0). Uses an autogenerated depth value, the same for all points passed.

Will add all positions. Missing radii will default to Size::AUTO. Missing colors will default to white.

pub fn flags(self, flags: PointCloudBatchFlags) -> Self

Adds (!) flags for this batch.

pub fn picking_object_id(self, picking_object_id: PickingLayerObjectId) -> Self

Sets the picking object id for the current batch.

pub fn push_additional_outline_mask_ids_for_range( self, range: Range<u32>, ids: OutlineMaskPreference ) -> Self

Pushes additional outline mask ids for a specific range of points. The range is relative to this batch.

Prefer the overall_outline_mask_ids setting to set the outline mask ids for the entire batch whenever possible!