Struct re_renderer::Size

#[repr(C)]
pub struct Size(pub f32);

A size of something in either world-units, screen-units, or unsized.

Implementation:

If positive, this is in scene units.

If negative, this is in points.

If NaN, auto-size it.

Resolved on-the-fly in shader code. See shader/utils/size.wgsl

Tuple Fields
0: f32
Implementations

impl Size

pub const AUTO: Self = _

Automatically sized, based on a view builder setting.

pub const AUTO_LARGE: Self = _

Like Size::AUTO, but larger by some small factor (~2).

pub fn new_scene(size: f32) -> Self

Creates a new size in scene units.

Values passed must be finite positive.

pub fn new_points(size: f32) -> Self

Creates a new size in ui point units.

Values passed must be finite positive.

pub fn is_auto(&self) -> bool

Returns true if the size is an automatically determined size (Self::AUTO or Self::AUTO_LARGE).

pub fn scene(&self) -> Option<f32>

Get the scene-size of this, if stored as a scene size.

pub fn points(&self) -> Option<f32>

Get the point size of this, if stored as a point size.