Struct re_renderer::texture_info::Texture2DBufferInfo

pub struct Texture2DBufferInfo {
    pub bytes_per_row_unpadded: u32,
    pub bytes_per_row_padded: u32,
    pub buffer_size_unpadded: BufferAddress,
    pub buffer_size_padded: BufferAddress,
}

Utility for dealing with buffers containing raw 2D texture data.

Fields
bytes_per_row_unpadded: u32

How many bytes per row contain actual data.

bytes_per_row_padded: u32

How many bytes per row are required to be allocated in total.

Padding bytes are always at the end of a row.

buffer_size_unpadded: BufferAddress

Size required for an unpadded buffer.

buffer_size_padded: BufferAddress

Size required for a padded buffer as it is read/written from/to the GPU.

Implementations

impl Texture2DBufferInfo

pub fn new(format: TextureFormat, extent: UVec2) -> Self

Retrieves 2D texture buffer info for a given format & texture size.

If a single buffer is not possible for all aspects of the texture format, all sizes will be zero.

pub fn num_rows(&self) -> u32

pub fn remove_padding<'a>(&self, buffer: &'a [u8]) -> Cow<'a, [u8]>

Removes the padding from a buffer containing gpu texture data.

The passed in buffer is to be expected to be exactly of size Texture2DBufferInfo::buffer_size_padded.

Note that if you’re passing in gpu data, there no alignment guarantees on the returned slice, do NOT convert it using [bytemuck]. Use Texture2DBufferInfo::remove_padding_and_convert instead.

pub fn remove_padding_and_convert<T: Pod>(&self, buffer: &[u8]) -> Vec<T>

Removes the padding from a buffer containing gpu texture data and remove convert to a given type.

The passed in buffer is to be expected to be exactly of size Texture2DBufferInfo::buffer_size_padded.

The unpadded row size is expected to be a multiple of the size of the target type. (Which means that, while uncommon, it technically doesn’t need to be as big as a block in the pixel - this can be useful for e.g. packing wide bitfields)