re_renderer
Type Alias RecommendedFileResolver

pub type RecommendedFileResolver = FileResolver<&'static MemFileSystem>;

The recommended FileResolver type for the current platform/target.

Aliased Type

struct RecommendedFileResolver {
    fs: &'static MemFileSystem,
    search_path: SearchPath,
}

Fields
fs: &'static MemFileSystem

A handle to the filesystem being used. Generally a OsFileSystem on native and a MemFileSystem on web and during tests.

search_path: SearchPath

The search path that we will go through when an import cannot be resolved neither as an absolute path or a relative one.