Module re_renderer::wgpu_resources

Wgpu resource pools are concerned with handling low level gpu resources efficiently.

They facilitate easy creation and avoidance of unnecessary gpu allocations.

This is in contrast to the crate::resource_managers which are concerned with higher level resources that arise from processing user provided data.

Re-exports

pub use bind_group_layout_pool::BindGroupLayoutDesc;

pub use bind_group_layout_pool::GpuBindGroupLayoutHandle;

pub use bind_group_layout_pool::GpuBindGroupLayoutPool;

pub use bind_group_pool::BindGroupDesc;

pub use bind_group_pool::BindGroupEntry;

pub use bind_group_pool::GpuBindGroup;

pub use bind_group_pool::GpuBindGroupHandle;

pub use bind_group_pool::GpuBindGroupPool;

pub use buffer_pool::BufferDesc;

pub use buffer_pool::GpuBuffer;

pub use buffer_pool::GpuBufferHandle;

pub use buffer_pool::GpuBufferPool;

pub use pipeline_layout_pool::GpuPipelineLayoutHandle;

pub use pipeline_layout_pool::GpuPipelineLayoutPool;

pub use pipeline_layout_pool::PipelineLayoutDesc;

pub use render_pipeline_pool::GpuRenderPipelineHandle;

pub use render_pipeline_pool::GpuRenderPipelinePool;

pub use render_pipeline_pool::RenderPipelineDesc;

pub use render_pipeline_pool::VertexBufferLayout;

pub use sampler_pool::GpuSamplerHandle;

pub use sampler_pool::GpuSamplerPool;

pub use sampler_pool::SamplerDesc;

pub use shader_module_pool::GpuShaderModuleHandle;

pub use shader_module_pool::GpuShaderModulePool;

pub use shader_module_pool::ShaderModuleDesc;

pub use texture_pool::GpuTexture;

pub use texture_pool::GpuTextureHandle;

pub use texture_pool::GpuTextureInternal;

pub use texture_pool::GpuTexturePool;

pub use texture_pool::TextureDesc;

pub use resource::PoolError;

Modules

bind_group_layout_pool 🔒

bind_group_pool 🔒

buffer_pool 🔒

dynamic_resource_pool 🔒

pipeline_layout_pool 🔒

render_pipeline_pool 🔒

resource 🔒

sampler_pool 🔒

shader_module_pool 🔒

static_resource_pool 🔒

texture_pool 🔒

Structs

Texture2DBufferInfo

Utility for dealing with buffers containing raw 2D texture data.

WgpuResourcePoolStatistics

WgpuResourcePools

Collection of all wgpu resource pools.