Struct re_renderer::wgpu_resources::dynamic_resource_pool::DynamicResourcePool

pub(super) struct DynamicResourcePool<Handle: Key, Desc: Debug, Res> {
    state: RwLock<DynamicResourcePoolProtectedState<Handle, Desc, Res>>,
    current_frame_index: u64,
    total_resource_size_in_bytes: AtomicU64,
}

Generic resource pool for all resources that have varying contents beyond their description.

Unlike in super::static_resource_pool::StaticResourcePool, a resource can not be uniquely identified by its description, as the same description can apply to several different resources.

Fields
state: RwLock<DynamicResourcePoolProtectedState<Handle, Desc, Res>>current_frame_index: u64total_resource_size_in_bytes: AtomicU64
Implementations

impl<Handle, Desc, Res> DynamicResourcePool<Handle, Desc, Res>
where Handle: Key, Desc: Clone + Eq + Hash + Debug + DynamicResourcesDesc,

pub fn alloc<F: FnOnce(&Desc) -> Res>( &self, desc: &Desc, creation_func: F ) -> Arc<DynamicResource<Handle, Desc, Res>>

pub fn get_from_handle( &self, handle: Handle ) -> Result<Arc<DynamicResource<Handle, Desc, Res>>, PoolError>

pub fn begin_frame( &mut self, frame_index: u64, on_destroy_resource: impl FnMut(&Res) )

pub fn num_resources(&self) -> usize

pub fn total_resource_size_in_bytes(&self) -> u64