Struct re_renderer::wgpu_resources::render_pipeline_pool::RenderPipelineDesc

pub struct RenderPipelineDesc {
    pub label: DebugLabel,
    pub pipeline_layout: GpuPipelineLayoutHandle,
    pub vertex_entrypoint: String,
    pub vertex_handle: GpuShaderModuleHandle,
    pub fragment_entrypoint: String,
    pub fragment_handle: GpuShaderModuleHandle,
    pub vertex_buffers: SmallVec<[VertexBufferLayout; 4]>,
    pub render_targets: SmallVec<[Option<ColorTargetState>; 4]>,
    pub primitive: PrimitiveState,
    pub depth_stencil: Option<DepthStencilState>,
    pub multisample: MultisampleState,
}

Renderpipeline descriptor, can be converted into [wgpu::RenderPipeline] (which isn’t hashable or comparable)

Fields
label: DebugLabel

Debug label of the pipeline. This will show up in graphics debuggers for easy identification.

pipeline_layout: GpuPipelineLayoutHandlevertex_entrypoint: Stringvertex_handle: GpuShaderModuleHandlefragment_entrypoint: Stringfragment_handle: GpuShaderModuleHandlevertex_buffers: SmallVec<[VertexBufferLayout; 4]>

The format of any vertex buffers used with this pipeline.

render_targets: SmallVec<[Option<ColorTargetState>; 4]>

The color state of the render targets.

primitive: PrimitiveState

The properties of the pipeline at the primitive assembly and rasterization level.

depth_stencil: Option<DepthStencilState>

The effect of draw calls on the depth and stencil aspects of the output target, if any.

multisample: MultisampleState

The multi-sampling properties of the pipeline.

Implementations

impl RenderPipelineDesc

fn create_render_pipeline( &self, device: &Device, pipeline_layouts: &GpuPipelineLayoutPool, shader_modules: &GpuShaderModulePool ) -> Result<RenderPipeline>