Struct re_renderer::wgpu_resources::static_resource_pool::StaticResourcePool

pub(super) struct StaticResourcePool<Handle: Key, Desc, Res> {
    resources: RwLock<SlotMap<Handle, StoredResource<Res>>>,
    lookup: RwLock<HashMap<Desc, Handle>>,
    pub current_frame_index: u64,
}

Generic resource pool for all resources that are fully described upon creation.

This implies, a resource is uniquely defined by its description. We call these resources “static” because they never change their content during rendering. However, the description may be respect to indirect changes which may to recreation of a resource. The prime example of this is shader reloading:

The resource is semantically the exact same despite having a different wgpu resource.

We do not want its handle to change.

Fields
resources: RwLock<SlotMap<Handle, StoredResource<Res>>>lookup: RwLock<HashMap<Desc, Handle>>current_frame_index: u64
Implementations

impl<Handle, Desc, Res> StaticResourcePool<Handle, Desc, Res>
where Handle: Key, Desc: Debug + Clone + Eq + Hash,

pub fn get_or_create<F: FnOnce(&Desc) -> Res>( &self, desc: &Desc, creation_func: F ) -> Handle

pub fn recreate_resources<F: FnMut(&Desc) -> Option<Res>>( &mut self, recreation_func: F )

pub fn resources(&self) -> StaticResourcePoolReadLockAccessor<'_, Handle, Res>

Locks the resource pool for resolving handles.

While it is locked, no new resources can be added.

pub fn take_resources( &mut self ) -> StaticResourcePoolMemMoveAccessor<Handle, Res>

Takes out all resources from the pool.

This is useful when the existing resources need to be accessed without taking a lock on the pool.

pub fn return_resources( &mut self, resources: StaticResourcePoolMemMoveAccessor<Handle, Res> )

Counter part to take_resources.

Logs an error if resources were added to the pool since take_resources was called.

pub fn num_resources(&self) -> usize