Struct re_renderer::wgpu_resources::texture_pool::GpuTexturePool

pub struct GpuTexturePool {
    pool: DynamicResourcePool<GpuTextureHandle, TextureDesc, GpuTextureInternal>,
}

Fields
pool: DynamicResourcePool<GpuTextureHandle, TextureDesc, GpuTextureInternal>
Implementations

impl GpuTexturePool

pub fn alloc( &self, device: &Device, desc: &TextureDesc ) -> Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>

Returns a reference-counted handle to a currently unused texture. Once ownership to the handle is given up, the texture may be reclaimed in future frames.

pub fn begin_frame(&mut self, frame_index: u64)

Called by RenderContext every frame. Updates statistics and may free unused textures.

pub fn get_from_handle( &self, handle: GpuTextureHandle ) -> Result<Arc<DynamicResource<GpuTextureHandle, TextureDesc, GpuTextureInternal>>, PoolError>

Method to retrieve a resource from a weak handle (used by super::GpuBindGroupPool)

pub fn num_resources(&self) -> usize

pub fn total_gpu_size_in_bytes(&self) -> u64