Module re_renderer::colormap

Enums

Colormap

Functions

colormap_cyan_to_yellow_srgb

Returns a gamma-space sRGB in 0-255 range. This is a perceptually uniform colormap which is robust to color blindness. It is especially suited for visualizing signed values. It interpolates from cyan to blue to dark gray to brass to yellow.

colormap_inferno_srgb

Returns sRGB polynomial approximation from Inferno color map, assuming t is normalized.

colormap_magma_srgb

Returns sRGB polynomial approximation from Magma color map, assuming t is normalized.

colormap_plasma_srgb

Returns sRGB polynomial approximation from Plasma color map, assuming t is normalized.

colormap_srgb

colormap_turbo_srgb

Returns sRGB polynomial approximation from Turbo color map, assuming t is normalized.

colormap_viridis_srgb

Returns sRGB polynomial approximation from Viridis color map, assuming t is normalized.

grayscale_srgb

Returns an sRGB gray value, assuming t is normalized.