Module re_renderer::colormap

Enums

Colormap

Functions

colormap_inferno_srgb

Returns sRGB polynomial approximation from Inferno color map, assuming t is normalized.

colormap_magma_srgb

Returns sRGB polynomial approximation from Magma color map, assuming t is normalized.

colormap_plasma_srgb

Returns sRGB polynomial approximation from Plasma color map, assuming t is normalized.

colormap_srgb

colormap_turbo_srgb

Returns sRGB polynomial approximation from Turbo color map, assuming t is normalized.

colormap_viridis_srgb

Returns sRGB polynomial approximation from Viridis color map, assuming t is normalized.

grayscale_srgb

Returns an sRGB gray value, assuming t is normalized.