re_renderer::error_handling::wgpu_core_error
Trait DedupableError

trait DedupableError:
    Sized
    + Error
    + 'static {
    // Provided methods
    fn hash<H: Hasher>(&self, state: &mut H) { ... }
    fn eq(&self, rhs: &(dyn Error + Send + Sync + 'static)) -> bool { ... }
}

An error with some extra deduplication logic baked in.

Implemented by default for all relevant wgpu error types, though it might be worth providing a more specialized implementation for errors that are too broad by nature: e.g. a shader error cannot by deduplicated simply using the shader path.

Provided Methods

fn hash<H: Hasher>(&self, state: &mut H)

fn eq(&self, rhs: &(dyn Error + Send + Sync + 'static)) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.