Macro re_renderer::include_file

macro_rules! include_file {
    ($path:expr $(,)?) => { ... };
}

A macro to read the contents of a file on disk, and resolve #import clauses as required.

If load_shaders_from_disk is disabled, this will behave like the standard include_str macro.

If load_shaders_from_disk is enabled, this will actually load the specified path through our FileServer, and keep watching for changes in the background (both the root file and whatever direct and indirect dependencies it may have through #import clauses).