re_renderer
Module renderer

Modules

compositor 🔒

debug_overlay 🔒

depth_cloud 🔒

Renderer that makes it easy to draw point clouds straight out of depth textures.

generic_skybox 🔒

gpu_data

lines 🔒

Line renderer for efficient rendering of many line(strips)

mesh_renderer 🔒

Mesh renderer.

point_cloud 🔒

Point renderer for efficient rendering of point clouds.

rectangles 🔒

Renderer that makes it easy to draw textured 2D rectangles with transparency

test_triangle 🔒

world_grid 🔒

Structs

ColormappedTexture

Describes a texture and how to map it to a color.

DebugOverlayDrawData

Debug overlay for quick & dirty display of texture contents.

DebugOverlayRenderer

DepthCloud

DepthCloudDrawData

DepthCloudRenderer

DepthClouds

GenericSkyboxDrawData

GpuMeshInstance

LineBatchInfo

Data that is valid for a batch of line strips.

LineDrawData

A line drawing operation. Encompasses several lines, each consisting of a list of positions. Expected to be recreated every frame.

LineStripFlags

Property flags for a line strip

MeshDrawData

PointCloudBatchFlags

Property flags for a point batch

PointCloudBatchInfo

Data that is valid for a batch of point cloud points.

PointCloudDrawData

A point cloud drawing operation. Expected to be recreated every frame.

RectangleDrawData

RectangleOptions

TestTriangleDrawData

TexturedRect

WorldGridConfiguration

Configuration for the world grid renderer.

WorldGridDrawData

Draw data for a world grid renderer.

WorldGridRenderer

Enums

ColorMapper

How to map the normalized .r component to a color.

DebugOverlayError

DrawError

GenericSkyboxType

The type of generic skybox to render.

LineDrawDataError

PointCloudDrawDataError

ShaderDecoding

Describes how the color information is encoded in the texture.

TextureFilterMag

Texture filter setting for magnification (a texel covers several pixels).

TextureFilterMin

Texture filter setting for minification (several texels fall to one pixel).

Traits

DrawData

GPU sided data used by a Renderer to draw things to the screen.

Renderer

A Renderer encapsulate the knowledge of how to render a certain kind of primitives.

Functions

screen_triangle_vertex_shader

Gets or creates a vertex shader module for drawing a screen filling triangle.