re_renderer::renderer
Module depth_cloud

Renderer that makes it easy to draw point clouds straight out of depth textures.

Textures are uploaded just-in-time, no caching.

Implementation details

Since there’s no widespread support for bindless textures, this requires one bind group and one draw call per texture. This is a pretty heavy shader though, so the overhead is minimal.

The vertex shader backprojects the depth texture using the user-specified intrinsics, and then behaves pretty much exactly like our point cloud renderer (see [point_cloud.rs]).

Modules

gpu_data 🔒

Structs

DepthCloud

DepthCloudDrawData

DepthCloudDrawInstance 🔒

DepthCloudRenderer

DepthClouds

Enums

DepthCloudDrawDataError