Module re_renderer::renderer::point_cloud

Point renderer for efficient rendering of point clouds.

How it works:

Points are rendered as quads and stenciled out by a fragment shader. Quad spanning happens in the vertex shader, uploaded are only the data for the actual points (no vertex buffer!).

Like with the super::lines::LineRenderer, we’re rendering as all quads in a single triangle list draw call. (Rationale for this can be found in the [lines.rs]’s documentation)

For WebGL compatibility, data is uploaded as textures. Color is stored in a separate srgb texture, meaning that srgb->linear conversion happens on texture load.

Modules

gpu_data

Structs

PointCloudBatch 🔒

Internal, ready to draw representation of PointCloudBatchInfo

PointCloudBatchFlags

Property flags for a point batch

PointCloudBatchInfo

Data that is valid for a batch of point cloud points.

PointCloudDrawData

A point cloud drawing operation. Expected to be recreated every frame.

PointCloudRenderer

Enums

PointCloudDrawDataError