re_renderer::renderer
Module rectangles

Renderer that makes it easy to draw textured 2D rectangles with transparency

Transparency: (TODO(andreas):) We’re not performing any sorting on transparency yet, so the transparent rectangles pretty much only work correctly when they are directly layered in front of another opaque rectangle. We do not disable depth write.

Implementation details: We assume the standard usecase are individual textured rectangles. Since we’re not allowed to bind many textures at once (no widespread bindless support!), we are forced to have individual bind groups per rectangle and thus a draw call per rectangle.

Modules

gpu_data 🔒

Structs

ColormappedTexture

Describes a texture and how to map it to a color.

RectangleDrawData

RectangleInstance 🔒

RectangleOptions

RectangleRenderer

TexturedRect

Enums

ColorMapper

How to map the normalized .r component to a color.

RectangleError

ShaderDecoding

Describes how the color information is encoded in the texture.

TextureFilterMag

Texture filter setting for magnification (a texel covers several pixels).

TextureFilterMin

Texture filter setting for minification (several texels fall to one pixel).