Trait re_renderer::renderer::Renderer

pub trait Renderer {
    type RendererDrawData: DrawData;

    // Required methods
    fn create_renderer<Fs: FileSystem>(
        shared_data: &SharedRendererData,
        pools: &mut WgpuResourcePools,
        device: &Device,
        resolver: &mut FileResolver<Fs>
    ) -> Self;
    fn draw<'a>(
        &self,
        pools: &'a WgpuResourcePools,
        phase: DrawPhase,
        pass: &mut RenderPass<'a>,
        draw_data: &'a Self::RendererDrawData
    ) -> Result<(), DrawError>;
    fn participated_phases() -> &'static [DrawPhase];
}

A Renderer encapsulate the knowledge of how to render a certain kind of primitives.

It is an immutable, long-lived datastructure that only holds onto resources that will be needed for each of its Renderer::draw invocations. Any data that might be different per specific Renderer::draw invocation is stored in DrawData.

Required Associated Types

type RendererDrawData: DrawData

Required Methods

fn create_renderer<Fs: FileSystem>( shared_data: &SharedRendererData, pools: &mut WgpuResourcePools, device: &Device, resolver: &mut FileResolver<Fs> ) -> Self

fn draw<'a>( &self, pools: &'a WgpuResourcePools, phase: DrawPhase, pass: &mut RenderPass<'a>, draw_data: &'a Self::RendererDrawData ) -> Result<(), DrawError>

Called once per phase given by Renderer::participated_phases.

fn participated_phases() -> &'static [DrawPhase]

Combination of flags indicating in which phases Renderer::draw should be called.