Module re_renderer::resource_managers

Resource managers are concerned with mapping (typically) higher level user data to their Gpu representation.

They facilitate fast & easy gpu upload and resource usage.

This is in contrast to the pools in crate::wgpu_resources which are exclusively concerned with low level gpu resources and their efficient allocation.

Modules

mesh_manager 🔒

resource_manager 🔒

texture_manager 🔒

Structs

GpuTexture2D

Handle to a 2D resource.

MeshManager

Texture2DCreationDesc

Data required to create a texture 2D resource.

TextureManager2D

Texture manager for 2D textures.

Enums

ResourceHandle

Handle to a resource that is stored in a resource manager.

ResourceLifeTime

ResourceManagerError

TextureCreationError

TextureManager2DError

Type Aliases

GpuMeshHandle