Type Alias re_renderer::RecommendedFileResolver

pub type RecommendedFileResolver = FileResolver<OsFileSystem>;

The recommended FileResolver type for the current platform/target.

Aliased Type

struct RecommendedFileResolver {
    fs: OsFileSystem,
    search_path: SearchPath,
}

Fields
fs: OsFileSystem

A handle to the filesystem being used. Generally a OsFileSystem on native and a MemFileSystem on web and during tests.

search_path: SearchPath

The search path that we will go through when an import cannot be resolved neither as an absolute path or a relative one.

Implementations

impl<Fs: FileSystem> FileResolver<Fs>

pub fn new(fs: Fs) -> Self

pub fn with_search_path(fs: Fs, search_path: SearchPath) -> Self

impl<Fs: FileSystem> FileResolver<Fs>

pub fn populate(&self, path: impl AsRef<Path>) -> Result<InterpolatedFile>

fn resolve_clause_path( &self, cwd: impl AsRef<Path>, path: impl AsRef<Path> ) -> Option<PathBuf>