Module re_renderer::view_builder

Structs

AutoSizeConfig

How Size::AUTO is interpreted.

TargetConfiguration

Basic configuration for a target view.

ViewBuilder

The highest level rendering block in re_renderer. Used to build up/collect various resources and then send them off for rendering of a single view.

ViewTargetSetup 🔒

Enums

OrthographicCameraMode

Configures the camera placement in the orthographic frustum, as well as the coordinate system convention.

Projection

How we project from 3D to 2D.

ViewBuilderError

Constants

AUTO_LINE_RADIUS 🔒

AUTO_POINT_RADIUS 🔒

Type Definitions

SharedViewBuilder

ViewBuilder that can be shared between threads.